add: 피드 작성 시 사진 s3업로드 및 삭제 시 s3 파일 기능 추가

+++ src/main/java/com/onehana/server_ilogu/service/AmazonS3Service.java
@@ -3,7 +3,16 @@ package com.onehana.server_ilogu.service;
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.PutObjectRequest;
-import com.onehana.server_ilogu.dto.S3FileDto;
+import com.onehana.server_ilogu.dto.BoardDto;
+import com.onehana.server_ilogu.dto.BoardImageDto;
+import com.onehana.server_ilogu.dto.response.BaseResponseStatus;
+import com.onehana.server_ilogu.entity.Board;
+import com.onehana.server_ilogu.entity.BoardImage;
+import com.onehana.server_ilogu.entity.User;
+import com.onehana.server_ilogu.exception.BaseException;
+import com.onehana.server_ilogu.repository.BoardImageRepository;
+import com.onehana.server_ilogu.repository.BoardRepository;
+import com.onehana.server_ilogu.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -14,6 +23,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.text.SimpleDateFormat;
 import java.util.*;
+import java.util.stream.Collectors;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -23,23 +33,25 @@ public class AmazonS3Service {
     @Value("${cloud.aws.s3.bucket}")
     private String bucket;
 
+    private final UserRepository userRepository;
+    private final BoardRepository boardRepository;
+    private final BoardImageRepository boardImageRepository;
     private final AmazonS3Client amazonS3Client;
 
-    public List<S3FileDto> uploadFiles(List<MultipartFile> multipartFiles) {
-        List<S3FileDto> s3files = new ArrayList<>();
-        String uploadFilePath = "temp" + "/" + getFolderName();
+    public List<BoardImageDto> uploadFiles(List<MultipartFile> files, BoardDto boardDto) {
+        List<BoardImageDto> s3files = new ArrayList<>();
+        String uploadFilePath = "board" + "/" + getFolderName();
 
-        for (MultipartFile multipartFile : multipartFiles) {
-            String originalFileName = multipartFile.getOriginalFilename();
+        for (MultipartFile file : files) {
+            String originalFileName = file.getOriginalFilename();
             String uploadFileName = getUuidFileName(originalFileName);
             String uploadFileUrl = "";
 
             ObjectMetadata objectMetadata = new ObjectMetadata();
-            objectMetadata.setContentLength(multipartFile.getSize());
-            objectMetadata.setContentType(multipartFile.getContentType());
+            objectMetadata.setContentLength(file.getSize());
+            objectMetadata.setContentType(file.getContentType());
 
-            try (InputStream inputStream = multipartFile.getInputStream()) {
-
+            try (InputStream inputStream = file.getInputStream()) {
                 String keyName = uploadFilePath + "/" + uploadFileName;
 
                 amazonS3Client.putObject(
@@ -49,42 +61,53 @@ public class AmazonS3Service {
 
             } catch (IOException e) {
                 e.printStackTrace();
-                log.error("Filed upload failed", e);
             }
 
             s3files.add(
-                    S3FileDto.builder()
+                    BoardImageDto.builder()
                             .originalFileName(originalFileName)
                             .uploadFileName(uploadFileName)
                             .uploadFilePath(uploadFilePath)
                             .uploadFileUrl(uploadFileUrl)
                             .build());
         }
+        Board board = boardRepository.findById(boardDto.getId())
+                .orElseThrow(() -> new BaseException(BaseResponseStatus.BOARD_NOT_FOUND));
+
+        List<BoardImage> boardImages = s3files.stream()
+                .map(dto -> BoardImage.from(dto, board))
+                .collect(Collectors.toList());
+
+        boardImageRepository.saveAll(boardImages);
         return s3files;
     }
 
-    public String deleteFile(String uploadFilePath, String uuidFileName) {
+    public void deleteFile(String email, Long boardId) {
+        User user = getUserOrException(email);
+        Board board = getBoardOrException(boardId);
 
-        String result = "success";
+        if (board.getUser() != user) {
+            throw new BaseException(BaseResponseStatus.INVALID_PERMISSION);
+        }
 
-        try {
-            String keyName = uploadFilePath + "/" + uuidFileName;
+        List<BoardImage> boardImages = boardImageRepository.findByBoardId(boardId).orElseThrow(() ->
+                new BaseException(BaseResponseStatus.BOARD_NOT_FOUND));
+
+        for (BoardImage boardImage : boardImages) {
+            String keyName = boardImage.getUploadFilePath() + "/" + boardImage.getUploadFileName();
             boolean isObjectExist = amazonS3Client.doesObjectExist(bucket, keyName);
+
             if (isObjectExist) {
                 amazonS3Client.deleteObject(bucket, keyName);
             } else {
-                result = "file not found";
+                throw new BaseException(BaseResponseStatus.IMAGE_NOT_FOUND);
             }
-        } catch (Exception e) {
-            log.debug("Delete File failed", e);
         }
-
-        return result;
     }
 
     public String getUuidFileName(String fileName) {
         String ext = fileName.substring(fileName.indexOf(".") + 1);
-        return UUID.randomUUID().toString() + "." + ext;
+        return UUID.randomUUID() + "." + ext;
     }
 
     private String getFolderName() {
@@ -92,5 +115,15 @@ public class AmazonS3Service {
         Date date = new Date();
         String str = sdf.format(date);
         return str.replace("-", "/");
+    }
+
+    private User getUserOrException(String email) {
+        return userRepository.findByEmail(email).orElseThrow(() ->
+                new BaseException(BaseResponseStatus.USER_NOT_FOUND));
+    }
+
+    private Board getBoardOrException(Long boardId) {
+        return boardRepository.findById(boardId).orElseThrow(() ->
+                new BaseException(BaseResponseStatus.BOARD_NOT_FOUND));
     }
 }
